+++ public/utils/config-equivalences.js
@@ -9,9 +9,9 @@ export const configEquivalences = {
     'Enable or disable the setup health check when opening the app.',
   'checks.fields':
     'Enable or disable the known fields health check when opening the app.',
-    'checks.metaFields': 
+    'checks.metaFields':
       'Change the default value of the Kibana metaField configuration',
-    'checks.timeFilter': 
+    'checks.timeFilter':
       'Change the default value of the Kibana timeFilter configuration',
   'extensions.pci': 'Enable or disable the PCI DSS tab on Overview and Agents.',
   'extensions.gdpr': 'Enable or disable the GDPR tab on Overview and Agents.',
@@ -64,7 +64,9 @@ export const configEquivalences = {
   'cron.statistics.interval': 'Define the frequency of task execution using cron schedule expressions.',
   'cron.statistics.index.name': 'Define the name of the index in which the documents are to be saved.',
   'cron.statistics.index.creation': 'Define the interval in which the index will be created.',
-  'alerts.sample.prefix': 'Define the index name prefix of sample alerts. It must match with the template used by the index pattern to avoid unknown fields in dashboards.'
+  'cron.statistics.index.shards': 'Define the number of shards to use for the statistics indices.',
+  'cron.statistics.index.replicas': 'Define the number of replicas to use for the statistics indices.',
+  'alerts.sample.prefix': 'Define the index name prefix of sample alerts. It must match with the template used by the index pattern to avoid unknown fields in dashboards.',
 };
 
 export const nameEquivalence = {
@@ -96,6 +98,8 @@ export const nameEquivalence = {
   'cron.statistics.interval': 'Interval',
   'cron.statistics.index.name': 'Index name',
   'cron.statistics.index.creation': 'Index creation',
+  'cron.statistics.index.shards': 'Index shards',
+  'cron.statistics.index.replicas': 'Index replicas',
   'alerts.sample.prefix': 'Sample alerts prefix',
 }
 
@@ -134,6 +138,8 @@ export const categoriesEquivalence = {
   'cron.statistics.interval': STATISTICS,
   'cron.statistics.index.name': STATISTICS,
   'cron.statistics.index.creation': STATISTICS,
+  'cron.statistics.index.shards': STATISTICS,
+  'cron.statistics.index.replicas': STATISTICS,
   'alerts.sample.prefix': GENERAL,
 }
 
@@ -198,5 +204,7 @@ export const formEquivalence = {
       ]
     }
   },
+  'cron.statistics.index.shards': { type: NUMBER },
+  'cron.statistics.index.replicas': { type: NUMBER },
   'alerts.sample.prefix': { type: TEXT },
 }