+++ util/constants.js
@@ -40,5 +40,9 @@ export const WAZUH_TIME_FILTER_DEFAULT = {
   to: 'now'
 };
 
+// Default number of shards and remplicas for indices
+export const WAZUH_INDEX_SHARDS = 2;
+export const WAZUH_INDEX_REPLICAS = 0;
+
 // Reserved ids for Users/Role mapping
 export const WAZUH_API_RESERVED_ID_LOWER_THAN = 100;