+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Sample data indices name should take index pattern in use [#2593](https://github.com/wazuh/wazuh-kibana-app/issues/2593) 
 - Added start option to macos Agents [#2653](https://github.com/wazuh/wazuh-kibana-app/pull/2653)
 
+### Changed
+
+- Statistics settings do not allow to configure primary shards and replicas [#2627](https://github.com/wazuh/wazuh-kibana-app/issues/2627)
+
 ## Wazuh v4.0.2 - Kibana v7.9.1, v7.9.3 - Revision 4011
 
 ### Added